Create main content, design, and logic of web app
I created the structure and main content of the "Page by Page" library
app with a big header for the name of the app and a book gallery showing
the books you have read in a card format.

The layout of the library app utilized CSS Grid and was designed with
a color palette that matches the vibes of a book/library.

For the logic of the web app, I wrote an object constructor for the
books. Each book is an instance of this object that contain details such
as the title, author, published date, genre, and a unique identifier
generated using the crypto.randomUUID function. The book object is then
stored in an array and is then displayed to the library app.

For testing the display and logic, I added five books of my choosing
using the addBookToLibrary function.