Add 'New Book' functionality with modal and form
I added a button that allows the user to add a 'New Book' in their
library. Clicking the button shows a modal containing a form that the
user can fill the details of the new book they would like to add. Once
the form is submitted, a new book card is displayed in their book
gallery. I also added two new details for the book: Number of pages and
Have they read the book?.

As for the script, I removed the displayToLibrary function and replaced
it with the createBookCard function. This function makes it so that the
displayed book cards are created programatically and ensures that the
code is highly maintainable. Additionally, the addBookToLibrary function
also utilizes the createBookCard function.

For the modal form, I added three eventListeners that would show the
modal, close the modal, and submit the form inside the modal. When the
form is submitted, it also clears the values inside of the input
elements.