Add 'Toggle Read Status' functionality
I added a new toggle switch button that replaces the old text that
displays if the user has already read the book or not. The toggle switch
is designed by placing the checkbox input inside the label and changing
their appearance through CSS. I also added a transition to animate the
switching of the toggle switch.

For the script, I programatically created the building blocks of the
toggle switch and append it as a child of the created book card. I also
created a prototype function for the Book object that updates the
hasReadBook property of the book object inside the myLibrary array.